+++ assets/Resources/Collision.h
@@ -31,8 +31,8 @@ public:
     void addCollider(const glm::ivec4 &boxCollider);
     void removeCollider(const glm::ivec4 &boxCollider);
 
-    void changePositionAbsolute(const glm::ivec2 &pos);
-    void changePositionRelative(const glm::ivec2 &pos);
+    void changePositionAbsolute(const glm::vec2 &pos);
+    void changePositionRelative(const glm::vec2 &pos);
 
 #ifdef SHOW_HIT_BOXES
     void showHitBox();
@@ -46,7 +46,7 @@ public:
     int collidersSize;
     glm::ivec4 *collisions;
 
-    glm::ivec2 position;
+    glm::vec2 position;
 
 private:
